+++ internal/auth.go
@@ -48,9 +48,19 @@ func (p Provider) Login(email, password string) (accessToken, refreshToken strin
 		return "", "", fmt.Errorf("failed to generate access-token: %w", err)
 	}
 
-	refreshToken, err = p.JWTProvider.GenerateRefreshToken(email)
+	refreshToken, jwtID, err := p.JWTProvider.GenerateRefreshToken(email)
 	if err != nil {
 		return "", "", fmt.Errorf("failed to generate refresh-token: %w", err)
+	}
+
+	_, err = p.Storage.CreateToken(storage.Token{
+		EMail:     email,
+		Token:     jwtID,
+		Type:      storage.TokenTypeRefresh,
+		CreatedAt: nowFunc(),
+	})
+	if err != nil {
+		return "", "", fmt.Errorf("failed to persist refresh-token: %w", err)
 	}
 
 	return accessToken, refreshToken, nil
@@ -63,7 +73,7 @@ func (p Provider) Login(email, password string) (accessToken, refreshToken strin
 func (p Provider) Refresh(refreshToken string) (newAccessToken, newRefreshToken string, err error) {
 	isValid, claims, err := p.JWTProvider.IsTokenValid(refreshToken)
 	if err != nil {
-		return "", "", ErrTokenNotParsable
+		return "", "", fmt.Errorf("%w: %s", ErrTokenNotParsable, err)
 	}
 
 	if !isValid {
@@ -73,6 +83,35 @@ func (p Provider) Refresh(refreshToken string) (newAccessToken, newRefreshToken 
 	email, ok := claims["email"].(string)
 	if !ok {
 		return "", "", errors.New("email claim is not parsable as string")
+	}
+
+	tokenID, ok := claims["jit"].(string)
+	if !ok {
+		return "", "", errors.New("jit claim is not parsable as string")
+	}
+
+	//TODO do Storage.TokensByEMailAndToken and Storage.DeleteToken in transaction
+	tokens, err := p.Storage.TokensByEMailAndToken(email, tokenID)
+	if err != nil {
+		return "", "", fmt.Errorf("failed to find refresh-tokens: %w", err)
+	}
+
+	var t *storage.Token
+	for _, token := range tokens {
+		if token.Type == storage.TokenTypeRefresh {
+			// TODO check lifetime
+			t = &token
+			break
+		}
+	}
+
+	if t == nil {
+		return "", "", ErrNoValidTokenFound
+	}
+
+	err = p.Storage.DeleteToken(t.ID)
+	if err != nil {
+		return "", "", fmt.Errorf("failed to delete refresh-token: %w", err)
 	}
 
 	u, err := p.Storage.User(email)
@@ -88,9 +127,19 @@ func (p Provider) Refresh(refreshToken string) (newAccessToken, newRefreshToken 
 		return "", "", fmt.Errorf("failed to generate access-token: %w", err)
 	}
 
-	newRefreshToken, err = p.JWTProvider.GenerateRefreshToken(email)
+	newRefreshToken, jwtID, err := p.JWTProvider.GenerateRefreshToken(email)
 	if err != nil {
 		return "", "", fmt.Errorf("failed to generate refresh-token: %w", err)
+	}
+
+	_, err = p.Storage.CreateToken(storage.Token{
+		EMail:     email,
+		Token:     jwtID,
+		Type:      storage.TokenTypeRefresh,
+		CreatedAt: nowFunc(),
+	})
+	if err != nil {
+		return "", "", fmt.Errorf("failed to persist refresh-token: %w", err)
 	}
 
 	return newAccessToken, newRefreshToken, nil
@@ -133,9 +182,10 @@ func (p Provider) CreatePasswordResetRequest(email string) error {
 // ResetPassword resets the password of the given account if the reset token is correct.
 // return ErrNoValidTokenFound no valid token could be found
 func (p *Provider) ResetPassword(email, resetToken, newPassword string) error {
+	//TODO do Storage.TokensByEMailAndToken and Storage.DeleteToken in transaction
 	tokens, err := p.Storage.TokensByEMailAndToken(email, resetToken)
 	if err != nil {
-		return fmt.Errorf("failed to find tokens: %w", err)
+		return fmt.Errorf("failed to find reset-tokens: %w", err)
 	}
 
 	var t *storage.Token