+++ internal/auth_test.go
@@ -26,7 +26,9 @@ func TestProvider_Login(t *testing.T) {
 		generateAccessToken       string
 		generateAccessTokenError  error
 		generateRefreshToken      string
+		generateRefreshTokenID    string
 		generateRefreshTokenError error
+		createTokenError          error
 		dbReturnError             error
 		dbReturnUser              storage.User
 	}{
@@ -46,22 +48,19 @@ func TestProvider_Login(t *testing.T) {
 					"myCustomClaim": "value",
 				},
 			},
-		},
-		{
+		}, {
 			name:          "User not found",
 			givenEMail:    "not@existing.user",
 			givenPassword: "password",
 			expectedError: ErrUserNotFound,
 			dbReturnError: storage.ErrUserNotFound,
-		},
-		{
+		}, {
 			name:          "Unexpected db error",
 			givenEMail:    "not@existing.user",
 			givenPassword: "password",
 			expectedError: errors.New("failed to find user with email \"not@existing.user\": unexpected error"),
 			dbReturnError: errors.New("unexpected error"),
-		},
-		{
+		}, {
 			name: "Failed to generate accessToken",
 			dbReturnUser: storage.User{
 				Password: []byte("$2a$12$1v7O.pNLqugJjcePyxvUj.GK37YoAbJvSW/9bULSRmq5C4SkoU2OO"),
@@ -74,8 +73,7 @@ func TestProvider_Login(t *testing.T) {
 			givenPassword:            "password",
 			generateAccessTokenError: errors.New("error 42"),
 			expectedError:            errors.New("failed to generate access-token: error 42"),
-		},
-		{
+		}, {
 			name: "Failed to generate refreshToken",
 			dbReturnUser: storage.User{
 				Password: []byte("$2a$12$1v7O.pNLqugJjcePyxvUj.GK37YoAbJvSW/9bULSRmq5C4SkoU2OO"),
@@ -88,8 +86,7 @@ func TestProvider_Login(t *testing.T) {
 			givenPassword:             "password",
 			generateRefreshTokenError: errors.New("error 42"),
 			expectedError:             errors.New("failed to generate refresh-token: error 42"),
-		},
-		{
+		}, {
 			name:          "Incorrect Password",
 			givenEMail:    "test@test.test",
 			givenPassword: "wrongPassword",
@@ -97,6 +94,22 @@ func TestProvider_Login(t *testing.T) {
 			dbReturnUser: storage.User{
 				Password: []byte("$2a$12$1v7O.pNLqugJjcePyxvUj.GK37YoAbJvSW/9bULSRmq5C4SkoU2OO"),
 				EMail:    "test@test.test",
+			},
+		}, {
+			name:                   "Error while CreateToken",
+			givenEMail:             "test@test.test",
+			givenPassword:          "password",
+			createTokenError:       errors.New("nope"),
+			generatorExpectedEMail: "test@test.test",
+			generateAccessToken:    "myJWT",
+			generateRefreshToken:   "myRefreshJWT",
+			expectedError:          errors.New("failed to persist refresh-token: nope"),
+			dbReturnUser: storage.User{
+				Password: []byte("$2a$12$1v7O.pNLqugJjcePyxvUj.GK37YoAbJvSW/9bULSRmq5C4SkoU2OO"),
+				EMail:    "test@test.test",
+				Claims: map[string]interface{}{
+					"myCustomClaim": "value",
+				},
 			},
 		},
 	}
@@ -113,6 +126,9 @@ func TestProvider_Login(t *testing.T) {
 						givenStorageEMail = email
 						return tt.dbReturnUser, tt.dbReturnError
 					},
+					CreateTokenFunc: func(t storage.Token) (int64, error) {
+						return 0, tt.createTokenError
+					},
 				},
 				JWTProvider: &JWTProviderMock{
 					GenerateAccessTokenFunc: func(email string, userClaims map[string]interface{}) (string, error) {
@@ -120,9 +136,9 @@ func TestProvider_Login(t *testing.T) {
 						givenGenerateAccessTokenUserClaims = userClaims
 						return tt.generateAccessToken, tt.generateAccessTokenError
 					},
-					GenerateRefreshTokenFunc: func(email string) (string, error) {
+					GenerateRefreshTokenFunc: func(email string) (string, string, error) {
 						givenGenerateRefreshTokenEMail = email
-						return tt.generateRefreshToken, tt.generateRefreshTokenError
+						return tt.generateRefreshToken, tt.generateRefreshTokenID, tt.generateRefreshTokenError
 					},
 				},
 			}
@@ -166,24 +182,31 @@ func TestProvider_Refresh(t *testing.T) {
 	bcryptCost = bcrypt.MinCost
 
 	tests := []struct {
-		name                      string
-		email                     string
-		givenRefreshToken         string
-		givenPassword             string
-		expectedError             error
-		expectedAccessToken       string
-		expectedRefreshToken      string
-		generatorExpectedEMail    string
-		generateAccessToken       string
-		generateAccessTokenError  error
-		generateRefreshToken      string
-		generateRefreshTokenError error
-		IsTokenValidIsValid       bool
-		IsTokenValidClaims        jwt.MapClaims
-		isTokenValidErr           error
-		IsTokenValidToken         string
-		dbReturnError             error
-		dbReturnUser              storage.User
+		name                            string
+		email                           string
+		givenRefreshToken               string
+		givenPassword                   string
+		expectedError                   error
+		expectedAccessToken             string
+		expectedRefreshToken            string
+		expectedTokenID                 int64
+		expectedJWTID                   string
+		generatorExpectedEMail          string
+		generateAccessToken             string
+		generateAccessTokenError        error
+		generateRefreshToken            string
+		generateRefreshTokenError       error
+		generateRefreshTokenID          string
+		tokensByEMailAndTokenFuncTokens []storage.Token
+		tokensByEMailAndTokenFuncErr    error
+		deleteTokenErr                  error
+		isTokenValidIsValid             bool
+		isTokenValidClaims              jwt.MapClaims
+		isTokenValidErr                 error
+		isTokenValidToken               string
+		createTokenErr                  error
+		dbReturnError                   error
+		dbReturnUser                    storage.User
 	}{
 		{
 			name:                   "Happycase",
@@ -193,40 +216,48 @@ func TestProvider_Refresh(t *testing.T) {
 			generatorExpectedEMail: "test@test.test",
 			generateAccessToken:    "myJWT",
 			generateRefreshToken:   "myRefreshJWT",
-			IsTokenValidIsValid:    true,
-			IsTokenValidClaims:     jwt.MapClaims{"email": "test@test.test"},
-			IsTokenValidToken:      "givenRefreshToken",
-			expectedAccessToken:    "myJWT",
-			expectedRefreshToken:   "myRefreshJWT",
+			isTokenValidIsValid:    true,
+			isTokenValidClaims:     jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			isTokenValidToken:      "givenRefreshToken",
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			expectedAccessToken:  "myJWT",
+			expectedRefreshToken: "myRefreshJWT",
+			expectedJWTID:        "jwt-id",
+			expectedTokenID:      1234,
 			dbReturnUser: storage.User{
 				EMail: "test@test.test",
 				Claims: map[string]interface{}{
 					"myCustomClaim": "value",
 				},
 			},
-		},
-		{
+		}, {
 			name:                "User not found",
 			email:               "not@existing.user",
 			givenRefreshToken:   "givenRefreshToken",
 			givenPassword:       "password",
-			IsTokenValidIsValid: true,
-			IsTokenValidClaims:  jwt.MapClaims{"email": "not@existing.user"},
-			IsTokenValidToken:   "givenRefreshToken",
-			expectedError:       ErrUserNotFound,
-			dbReturnError:       storage.ErrUserNotFound,
-		},
-		{
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "not@existing.user", "jit": "jwt-id"},
+			isTokenValidToken:   "givenRefreshToken",
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			expectedError: ErrUserNotFound,
+			dbReturnError: storage.ErrUserNotFound,
+		}, {
 			name:                "Unexpected db error",
 			email:               "test@test.test",
 			givenRefreshToken:   "givenRefreshToken",
 			givenPassword:       "password",
-			IsTokenValidIsValid: true,
-			IsTokenValidClaims:  jwt.MapClaims{"email": "test@test.test"},
-			expectedError:       errors.New("failed to find user with email \"test@test.test\": unexpected error"),
-			dbReturnError:       errors.New("unexpected error"),
-		},
-		{
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			expectedError: errors.New("failed to find user with email \"test@test.test\": unexpected error"),
+			dbReturnError: errors.New("unexpected error"),
+		}, {
 			name:  "Failed to generate accessToken",
 			email: "test@test.test",
 			dbReturnUser: storage.User{
@@ -235,14 +266,16 @@ func TestProvider_Refresh(t *testing.T) {
 					"myCustomClaim": "value",
 				},
 			},
-			givenRefreshToken:        "not@existing.user",
-			givenPassword:            "password",
-			IsTokenValidIsValid:      true,
-			IsTokenValidClaims:       jwt.MapClaims{"email": "test@test.test"},
+			givenRefreshToken:   "not@existing.user",
+			givenPassword:       "password",
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
 			generateAccessTokenError: errors.New("error 42"),
 			expectedError:            errors.New("failed to generate access-token: error 42"),
-		},
-		{
+		}, {
 			name:  "Failed to generate refreshToken",
 			email: "test@test.test",
 			dbReturnUser: storage.User{
@@ -251,44 +284,91 @@ func TestProvider_Refresh(t *testing.T) {
 					"myCustomClaim": "value",
 				},
 			},
-			givenRefreshToken:         "not@existing.user",
-			givenPassword:             "password",
-			IsTokenValidIsValid:       true,
-			IsTokenValidClaims:        jwt.MapClaims{"email": "test@test.test"},
+			givenRefreshToken:   "not@existing.user",
+			givenPassword:       "password",
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
 			generateRefreshTokenError: errors.New("error 42"),
 			expectedError:             errors.New("failed to generate refresh-token: error 42"),
-		},
-		{
+		}, {
 			name:              "Token not parsable",
 			givenRefreshToken: "test@test.test",
 			givenPassword:     "wrongPassword",
-			isTokenValidErr:   errors.New("failed to parse token"),
-			expectedError:     ErrTokenNotParsable,
+			isTokenValidErr:   errors.New("given token is not parsable"),
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			expectedError: errors.New("given token is not parsable: given token is not parsable"),
 			dbReturnUser: storage.User{
 				EMail: "test@test.test",
 			},
-		},
-		{
+		}, {
 			name:                "Token not valid",
 			givenRefreshToken:   "test@test.test",
 			givenPassword:       "wrongPassword",
-			IsTokenValidIsValid: false,
-			IsTokenValidClaims:  jwt.MapClaims{"email": "test@test.test"},
+			isTokenValidIsValid: false,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test"},
 			expectedError:       ErrInvalidToken,
 			dbReturnUser: storage.User{
 				EMail: "test@test.test",
 			},
-		},
-		{
+		}, {
 			name:                "Token email claim is not a string",
 			givenRefreshToken:   "test@test.test",
 			givenPassword:       "wrongPassword",
-			IsTokenValidIsValid: true,
-			IsTokenValidClaims:  jwt.MapClaims{"email": 546544461176176},
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": 546544461176176, "jit": "jwt-id"},
 			expectedError:       errors.New("email claim is not parsable as string"),
 			dbReturnUser: storage.User{
 				EMail: "test@test.test",
 			},
+		}, {
+			name:                "Token jit claim is not present",
+			givenRefreshToken:   "test@test.test",
+			givenPassword:       "wrongPassword",
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": 123456},
+			expectedError:       errors.New("jit claim is not parsable as string"),
+			dbReturnUser: storage.User{
+				EMail: "test@test.test",
+			},
+		}, {
+			name:                            "No valid token found",
+			email:                           "test@test.test",
+			isTokenValidIsValid:             true,
+			isTokenValidClaims:              jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{},
+			expectedError:                   ErrNoValidTokenFound,
+		}, {
+			name:                         "Error while TokensByEMailAndToken",
+			email:                        "test@test.test",
+			isTokenValidIsValid:          true,
+			isTokenValidClaims:           jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncErr: errors.New("nope"),
+			expectedError:                errors.New("failed to find refresh-tokens: nope"),
+		}, {
+			name:                "Error while DeleteToken",
+			email:               "test@test.test",
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			deleteTokenErr: errors.New("nope"),
+			expectedError:  errors.New("failed to delete refresh-token: nope"),
+		}, {
+			name:                "Error while CreateToken",
+			email:               "test@test.test",
+			isTokenValidIsValid: true,
+			isTokenValidClaims:  jwt.MapClaims{"email": "test@test.test", "jit": "jwt-id"},
+			tokensByEMailAndTokenFuncTokens: []storage.Token{
+				{ID: 1234, EMail: "test.test@test.de", Type: storage.TokenTypeRefresh},
+			},
+			createTokenErr: errors.New("nope"),
+			expectedError:  errors.New("failed to persist refresh-token: nope"),
 		},
 	}
 
@@ -299,11 +379,26 @@ func TestProvider_Refresh(t *testing.T) {
 			var givenGenerateAccessTokenEMail string
 			var givenGenerateAccessTokenUserClaims map[string]interface{}
 			var givenIsTokenValidToken string
+			var givenTokensByEMailAndTokenEMail string
+			var givenTokensByEMailAndTokenToken string
+			var givenDeleteTokenID int64
 			toTest := Provider{
 				Storage: &StorageMock{
 					UserFunc: func(email string) (storage.User, error) {
 						givenStorageEMail = email
 						return tt.dbReturnUser, tt.dbReturnError
+					},
+					TokensByEMailAndTokenFunc: func(email string, token string) ([]storage.Token, error) {
+						givenTokensByEMailAndTokenEMail = email
+						givenTokensByEMailAndTokenToken = token
+						return tt.tokensByEMailAndTokenFuncTokens, tt.tokensByEMailAndTokenFuncErr
+					},
+					DeleteTokenFunc: func(id int64) error {
+						givenDeleteTokenID = id
+						return tt.deleteTokenErr
+					},
+					CreateTokenFunc: func(t storage.Token) (int64, error) {
+						return 0, tt.createTokenErr
 					},
 				},
 				JWTProvider: &JWTProviderMock{
@@ -312,13 +407,13 @@ func TestProvider_Refresh(t *testing.T) {
 						givenGenerateAccessTokenUserClaims = userClaims
 						return tt.generateAccessToken, tt.generateAccessTokenError
 					},
-					GenerateRefreshTokenFunc: func(email string) (string, error) {
+					GenerateRefreshTokenFunc: func(email string) (string, string, error) {
 						givenGenerateRefreshTokenEMail = email
-						return tt.generateRefreshToken, tt.generateRefreshTokenError
+						return tt.generateRefreshToken, tt.generateRefreshTokenID, tt.generateRefreshTokenError
 					},
 					IsTokenValidFunc: func(token string) (bool, jwt.MapClaims, error) {
 						givenIsTokenValidToken = token
-						return tt.IsTokenValidIsValid, tt.IsTokenValidClaims, tt.isTokenValidErr
+						return tt.isTokenValidIsValid, tt.isTokenValidClaims, tt.isTokenValidErr
 					},
 				},
 			}
@@ -343,7 +438,7 @@ func TestProvider_Refresh(t *testing.T) {
 			}
 
 			if givenGenerateAccessTokenEMail != tt.generatorExpectedEMail {
-				t.Errorf("Generator.GenerateAccessToken email ist not as expected: \nExpected:%s\nGiven:%s", tt.givenRefreshToken, givenGenerateAccessTokenEMail)
+				t.Errorf("Generator.GenerateAccessToken email is not as expected: \nExpected:%s\nGiven:%s", tt.givenRefreshToken, givenGenerateAccessTokenEMail)
 			}
 
 			if !reflect.DeepEqual(givenGenerateAccessTokenUserClaims, tt.dbReturnUser.Claims) {
@@ -351,11 +446,23 @@ func TestProvider_Refresh(t *testing.T) {
 			}
 
 			if givenGenerateRefreshTokenEMail != tt.generatorExpectedEMail {
-				t.Errorf("Generator.GenerateRefreshToken email ist not as expected: \nExpected:%s\nGiven:%s", tt.generatorExpectedEMail, givenGenerateRefreshTokenEMail)
+				t.Errorf("Generator.GenerateRefreshToken email is not as expected: \nExpected:%s\nGiven:%s", tt.generatorExpectedEMail, givenGenerateRefreshTokenEMail)
 			}
 
-			if givenIsTokenValidToken != tt.IsTokenValidToken {
-				t.Errorf("Generator.IsTokenValid token ist not as expected: \nExpected:%s\nGiven:%s", tt.IsTokenValidToken, givenIsTokenValidToken)
+			if givenIsTokenValidToken != tt.isTokenValidToken {
+				t.Errorf("Generator.IsTokenValid token is not as expected: \nExpected:%s\nGiven:%s", tt.isTokenValidToken, givenIsTokenValidToken)
+			}
+
+			if givenTokensByEMailAndTokenEMail != tt.email {
+				t.Errorf("Storage.TokensByEMailAndToken email is not as expected.\nExpected:%s\nGiven:%s", tt.email, givenTokensByEMailAndTokenEMail)
+			}
+
+			if tt.expectedJWTID != givenTokensByEMailAndTokenToken {
+				t.Errorf("Storage.TokensByEMailAndToken token is not as expected.\nExpected:%q\nGiven:%q", tt.expectedJWTID, givenTokensByEMailAndTokenToken)
+			}
+
+			if givenDeleteTokenID != tt.expectedTokenID {
+				t.Errorf("Storage.DeleteToken id is not as expected.\nExpected:%d\nGiven:%d", tt.expectedTokenID, givenDeleteTokenID)
 			}
 		})
 	}
@@ -531,7 +638,7 @@ func TestProvider_ResetPassword(t *testing.T) {
 			givenNewPassword: "newPassword",
 			givenResetToken:  "resetToken",
 			givenEMail:       "email",
-			expectedError:    errors.New("failed to find tokens: unexpected error"),
+			expectedError:    errors.New("failed to find reset-tokens: unexpected error"),
 			dbToken:          []storage.Token{},
 			dbTokenError:     errors.New("unexpected error"),
 		},