+++ internal/jwt/generate.go
@@ -30,7 +30,7 @@ func (p Provider) GenerateAccessToken(email string, userClaims map[string]interf
 	// standard claims by https://tools.ietf.org/html/rfc7519#section-4.1
 	claims["aud"] = p.privateClaims.audience      //Audience
 	claims["exp"] = now.Add(p.jwtLifetime).Unix() //ExpiresAt
-	claims["jit"] = jwtID                         //Id
+	claims["jit"] = jwtID.String()                //Id
 	claims["iat"] = now.Unix()                    //IssuedAt
 	claims["iss"] = p.privateClaims.issuer        //Issuer
 	claims["nbf"] = now.Unix()                    //NotBefore
@@ -39,7 +39,7 @@ func (p Provider) GenerateAccessToken(email string, userClaims map[string]interf
 	// public claims by https://www.iana.org/assignments/jwt/jwt.xhtml#claims
 	claims["email"] = email // Preferred e-mail address
 
-	token, err := jwt.NewWithClaims(jwt.SigningMethodES512, claims).SignedString(p.privateKey)
+	token, err := jwt.NewWithClaims(p.signingMethod, claims).SignedString(p.privateKey)
 	if err != nil {
 		return "", fmt.Errorf("failed to sign access-token: %w", err)
 	}
@@ -48,11 +48,11 @@ func (p Provider) GenerateAccessToken(email string, userClaims map[string]interf
 }
 
 // GenerateRefreshToken generates a valid refresh-jwt based on the Provider.privateKey. The jwt is issued to the given email.
-func (p Provider) GenerateRefreshToken(email string) (string, error) {
+func (p Provider) GenerateRefreshToken(email string) (string, string, error) {
 	now := timeNow()
 	jwtID, err := uuidNewRandom()
 	if err != nil {
-		return "", fmt.Errorf("failed to generate jwt-id: %w", err)
+		return "", "", fmt.Errorf("failed to generate jwt-id: %w", err)
 	}
 
 	claims := jwt.MapClaims{}
@@ -60,7 +60,7 @@ func (p Provider) GenerateRefreshToken(email string) (string, error) {
 	// standard claims by https://tools.ietf.org/html/rfc7519#section-4.1
 	claims["aud"] = p.privateClaims.audience             //Audience
 	claims["exp"] = now.Add(refreshTokenLifetime).Unix() //ExpiresAt
-	claims["jit"] = jwtID                                //Id
+	claims["jit"] = jwtID.String()                       //Id
 	claims["iat"] = now.Unix()                           //IssuedAt
 	claims["iss"] = p.privateClaims.issuer               //Issuer
 	claims["nbf"] = now.Unix()                           //NotBefore
@@ -69,12 +69,10 @@ func (p Provider) GenerateRefreshToken(email string) (string, error) {
 	// public claims by https://www.iana.org/assignments/jwt/jwt.xhtml#claims
 	claims["email"] = email // Preferred e-mail address
 
-	claims["refresh"] = true // Identify as refresh-token
-
-	token, err := jwt.NewWithClaims(jwt.SigningMethodES512, claims).SignedString(p.privateKey)
+	token, err := jwt.NewWithClaims(p.signingMethod, claims).SignedString(p.privateKey)
 	if err != nil {
-		return "", fmt.Errorf("failed to sign refresh-token: %w", err)
+		return "", "", fmt.Errorf("failed to sign refresh-token: %w", err)
 	}
 
-	return token, nil
+	return token, jwtID.String(), nil
 }