+++ internal/jwt/validate.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"github.com/dgrijalva/jwt-go"
-	"reflect"
 )
 
 var parseFunc = jwt.ParseWithClaims
@@ -13,14 +12,7 @@ var parseFunc = jwt.ParseWithClaims
 // isValid indicator, token-claims (when token is valid) and an error when present
 // return
 func (p Provider) IsTokenValid(tokenAsString string) (isValid bool, claims jwt.MapClaims, err error) {
-	token, err := parseFunc(tokenAsString, &claims, func(token *jwt.Token) (interface{}, error) {
-		tokenSigningMethod := reflect.TypeOf(token.Method)
-		if tokenSigningMethod != p.signingMethod {
-			return nil, fmt.Errorf("unexpected signing method %q", tokenSigningMethod)
-		}
-
-		return p.privateKey.PublicKey, nil
-	})
+	token, err := parseFunc(tokenAsString, &claims, checkSigningMethodKeyFunc(p.signingMethod, &p.privateKey.PublicKey))
 	if err != nil {
 		return false, nil, fmt.Errorf("failed to parse token: %w", err)
 	}