+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## unreleased
 - [[#3] secure admin api password config](https://github.com/leberKleber/simple-jwt-provider/issues/10)
 - [[#9] make jwt lifetime configurable](https://github.com/leberKleber/simple-jwt-provider/issues/9)
+- [[#6] refresh mechanism](https://github.com/leberKleber/simple-jwt-provider/issues/6)
 ## v1.0.0
 - password-reset flow via mail
 - admin api CRUD